autoforcad/ansible/roles/forcad-setup/files/checkers/requirements.txt
checklib==0.7.0
requests==2.31.0
pwntools==4.12.0
beautifulsoup4==4.14.2